Hotel Overcast & Redone Christmas variant
## Map Variants
- Hotel Snowver cast is now a brighter & snowier version of hotel overcast. Make sure you bring a winter coat for this one!

## Changelog
- Some blocks adjusted to avoid PGM overwriting them with barriers
- New building added to the apartment side:

It was originally too safe to bridge to the building due to the blocking of the bus stop area by the apartments. To combat this the bus stop has been replaced with a brand new apartment. This make any bridging to the apartment roof visible from spawn and should give bow

While it's not a perfect solution I'd rather avoid shortening the apartments to the point where they become a single storage as it'd detract from the look but I'm open to other suggestions in the future.